+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 ## [Unreleased][]
 
+### Bugfix
+- [#117](https://github.com/GDATASoftwareAG/robotframework-flaui/issues/117) : Combobox will not restore into Collapsed state after calling 4 related keywords. 
+
 ## [Release][2.0.13] [2.0.13][2.0.12-2.0.13] - 2023-07-25
 
 ### Bugfix